+++ src/app/core/services/configuration.repository.ts
@@ -3,8 +3,8 @@ import {HttpClient} from "@angular/common/http";
 import {Observable} from "rxjs";
 import {ConfigModel} from "../models/config.model";
 import {Inject, Injectable} from "@angular/core";
-import {ENV_TOKEN} from "../../app.module";
 import {Environment} from "../models/environment";
+import {ENV_TOKEN} from "../tokens";
 
 @Injectable({providedIn: "root"})
 export class ConfigurationRepository extends BaseRepository {