+++ src/app/product/components/product-list-page/product-list/product-list.component.spec.ts
@@ -1,8 +1,9 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ProductListComponent } from './product-list.component';
+import {CategoryEnum, ProductModel} from "../../../models/product.model";
 
-describe('ProductListComponentComponent', () => {
+describe('ProductListComponent', () => {
   let component: ProductListComponent;
   let fixture: ComponentFixture<ProductListComponent>;
 
@@ -21,5 +22,24 @@ describe('ProductListComponentComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should emit event on product buying', () => {
+    // Arrange
+    component.addItemToCart = jasmine.createSpyObj(['emit']);
+    const productModel: ProductModel = {
+      id: 1,
+      isAvailable: true,
+      name: 'Test',
+      description: 'Test',
+      category: CategoryEnum.Laptops,
+      price: 10.99,
+    };
+
+    // Act
+    component.onBuy(productModel);
+
+    // Assert
+    expect(component.addItemToCart.emit).toHaveBeenCalledWith(productModel);
   });
 });