+++ src/app/product/product-routing.module.ts
@@ -3,13 +3,11 @@ import {NgModule} from "@angular/core";
 import {ProductListPageComponent} from "./components/product-list-page/product-list-page.component";
 import {ProductViewPageComponent} from "./components/product-view-page/product-view-page.component";
 import {storeProductResolver} from "../shared/guards/product.resolver";
-import {ProductComponent} from "./components/product/product.component";
 import {fetchProduct} from "./store/product.actions";
 
 const routes: Routes = [
   {
     path: 'products',
-    component: ProductComponent,
     children: [
       {
         path: ':productID',