+++ src/app/shared/directives/highlight2.directive.spec.ts
@@ -1,5 +1,0 @@
-import { Highlight2Directive } from './highlight2.directive';
-
-describe('Highlight2Directive', () => {
-
-});